--- example/src/utils/consts.ts
@@ -1,2 +1,0 @@
-export const WIA_KEYTAG = "WIA_KEYTAG";
-export const DPOP_KEYTAG = "DPOP_KEYTAG";
