+++ example/src/utils/integrity.ts
